requirements.txt
# Install CPU-only torch separately using the official PyTorch CPU index URL before
# running `pip install -r requirements.txt`.
# Example:
# pip install torch==2.3.1 --index-url https://download.pytorch.org/whl/cpu

pdfplumber==0.10.3
beautifulsoup4==4.12.3
lxml==5.2.1
tiktoken==0.7.0
openai==1.35.0
chromadb==0.5.3
tenacity==8.3.0

tqdm==4.66.4
python-dotenv==1.0.1
numpy==1.26.4


torch==2.3.1




onnxruntime==1.20.0
optimum[onnxruntime]==1.20.0
transformers==4.41.2
tokenizers==0.19.1
numpy==1.26.4
